tg_bot: Reflect telegram changes to move from int32 to int64
Note: Needs DB Alteration

ALTER TABLE antiflood ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE gbans ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE gkicks ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE user_report_settings ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE userinfo ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE userbio ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE users ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE warns ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE chat_members ALTER COLUMN priv_chat_id TYPE BIGINT;
ALTER TABLE chat_members ALTER COLUMN user TYPE BIGINT;
ALTER TABLE chat_members ALTER COLUMN "user" TYPE BIGINT;

+++ tg_bot/modules/sql/users_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, Integer, UnicodeText, String, ForeignKey, UniqueConstraint, func
+from sqlalchemy import Column, BigInteger, Integer, UnicodeText, String, ForeignKey, UniqueConstraint, func
 
 from tg_bot import dispatcher
 from tg_bot.modules.sql import BASE, SESSION
@@ -8,7 +8,7 @@ from tg_bot.modules.sql import BASE, SESSION
 
 class Users(BASE):
     __tablename__ = "users"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     username = Column(UnicodeText)
 
     def __init__(self, user_id, username=None):
@@ -34,14 +34,14 @@ class Chats(BASE):
 
 class ChatMembers(BASE):
     __tablename__ = "chat_members"
-    priv_chat_id = Column(Integer, primary_key=True)
+    priv_chat_id = Column(BigInteger, primary_key=True)
     # NOTE: Use dual primary key instead of private primary key?
     chat = Column(String(14),
                   ForeignKey("chats.chat_id",
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
                   nullable=False)
-    user = Column(Integer,
+    user = Column(BigInteger,
                   ForeignKey("users.user_id",
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
